move debug binary main into crates/spyglass/binaries/debug
