+++ crates/spyglass/Cargo.toml
@@ -96,4 +96,4 @@ path = "src/main.rs"
 
 [[bin]]
 name = "spyglass-debug"
-path = "src/debug/main.rs"
+path = "binaries/debug/src/main.rs"
